+++ project.nix
@@ -5,7 +5,7 @@ let
 in
 nedryland.mkProject rec {
   name = "firm";
-  version = "2.0.3";
+  version = "4.0.0";
   baseExtensions = [
     ./extensions/nedryland/function.nix
     ./extensions/nedryland/runtime.nix