+++ services/lomax/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2022-02-14
+
 ### Added
 - Support for expansion of `{hostname}` to the real hostname in the
   `certificate_alt_names` config setting.